Use binary pickle to send/recv data
